fix: 🐛 解决字段模板中Oracle主键自增问题

+++ server/FieldTemplateServer.py
@@ -79,15 +79,24 @@ class FieldTemplate:
         return json.dumps(widget_type_property)
 
     def get_form_sql(self, form_id):
-        return '''DELETE FROM ea_form WHERE id = '{}';  # 删除模板
+        if self.db_type == "MYSQL":
+            return '''DELETE FROM ea_form WHERE id = '{}';  # 删除模板
 DELETE FROM ea_form_field WHERE form_id = '{}'; # 删除模板字段
+\n'''.format(form_id, form_id)
+        else:
+            return '''DELETE FROM ea_form WHERE id = '{}';
+DELETE FROM ea_form_field WHERE form_id = '{}';
 \n'''.format(form_id, form_id)
 
     def get_insert_field_sql(self, field, form_id, field_sort_number):
-        return "INSERT INTO ea_form_field (name, code, form_id, widget_type, sort_number, required, volume_primary_key, input_manually, tenant_id, created_date, last_modified_date, created_by, last_modified_by, widget_type_property) " \
+        if self.db_type == "MYSQL":
+            return "INSERT INTO ea_form_field (name, code, form_id, widget_type, sort_number, required, volume_primary_key, input_manually, tenant_id, created_date, last_modified_date, created_by, last_modified_by, widget_type_property) " \
                "VALUES ('{}', '{}', '{}', '{}', {}, {}, {}, {}, '-1', {}, {},  '-1', '-1', '{}');\n" \
             .format(field["字段名称*"], field["字段编码*"], form_id, self.get_field_widget_type(field["类型*"]), field_sort_number, self.get_boolean(field["是否必填*"]), self.get_boolean(field["是否成册主键"]), self.get_boolean(field["是否支持编辑*"]), self.get_now_time(), self.get_now_time(), str(self.get_field_widget_type_property(field)).replace(" ", ""))
-
+        else:
+            return "INSERT INTO ea_form_field (id, name, code, form_id, widget_type, sort_number, required, volume_primary_key, input_manually, tenant_id, created_date, last_modified_date, created_by, last_modified_by, widget_type_property) " \
+                   "VALUES ({}, '{}', '{}', '{}', '{}', {}, {}, {}, {}, '-1', {}, {},  '-1', '-1', '{}');\n" \
+                .format("seq_on_common_tabs.nextval", field["字段名称*"], field["字段编码*"], form_id, self.get_field_widget_type(field["类型*"]), field_sort_number, self.get_boolean(field["是否必填*"]), self.get_boolean(field["是否成册主键"]), self.get_boolean(field["是否支持编辑*"]), self.get_now_time(), self.get_now_time(), str(self.get_field_widget_type_property(field)).replace(" ", ""))
     def get_now_time(self):
         if self.db_type == "MYSQL":
             return "NOW()"
@@ -125,7 +134,8 @@ DELETE FROM ea_form_field WHERE form_id = '{}'; # 删除模板字段
         form_id = 10
 
         for form, fields in self.format_content():
-            sql = sql + "# {}字段模板\n".format(form[1])
+            if self.db_type == "MYSQL":
+                sql = sql + "# {}字段模板\n".format(form[1])
             sql = sql + self.get_form_sql(form_id)
             sql = sql + self.get_insert_form_sql(form_id, form[0], form[1], self.get_business_type(form[2]))
             field_sort_number = 0
